backend/services/ai_agent.py
"""
AI Content Generation Agent using Strands SDK
"""
from strands import Agent, tool
from core.config import settings
import httpx
from typing import List, Optional


@tool
async def generate_social_caption(
    content_type: str,
    topic: str,
    platform: str,
    tone: str = "professional"
) -> str:
    """Generate engaging social media captions.
    
    Args:
        content_type: Type of content (image, video, link, text)
        topic: Main topic or theme
        platform: Target platform (twitter, linkedin, instagram, facebook)
        tone: Desired tone (professional, casual, humorous, inspirational)
    
    Returns:
        Generated caption text
    """
    # This will be implemented with actual Strands agent logic
    # For now, return a placeholder
    return f"Generated {tone} caption for {content_type} about {topic} on {platform}"


@tool
async def suggest_hashtags(
    content: str,
    platform: str,
    max_hashtags: int = 10
) -> List[str]:
    """Suggest relevant hashtags based on content.
    
    Args:
        content: The post content to analyze
        platform: Target platform (twitter, linkedin, instagram, facebook)
        max_hashtags: Maximum number of hashtags to suggest
    
    Returns:
        List of suggested hashtags
    """
    # This will be implemented with actual hashtag analysis
    # For now, return placeholder hashtags
    return [f"#{platform}", "#socialmedia", "#content", "#marketing"]


@tool
async def generate_image_with_nano_banana(
    prompt: str,
    style: str = "realistic",
    dimensions: str = "1024x1024"
) -> str:
    """Generate images using Google Nano Banana.
    
    Args:
        prompt: Description of the image to generate
        style: Image style (realistic, artistic, cartoon, etc.)
        dimensions: Image dimensions (1024x1024, 1920x1080, etc.)
    
    Returns:
        URL or path to the generated image
    """
    # This will be implemented with Google Nano Banana API
    # For now, return a placeholder URL
    return f"https://example.com/generated-image-{style}-{dimensions}.jpg"


@tool
async def generate_content_ideas(
    topic: str,
    audience: str,
    platform: str,
    count: int = 5
) -> List[str]:
    """Generate content ideas for social media.
    
    Args:
        topic: Main topic or theme
        audience: Target audience description
        platform: Target platform
        count: Number of ideas to generate
    
    Returns:
        List of content ideas
    """
    # This will be implemented with actual content generation
    # For now, return placeholder ideas
    return [
        f"Share insights about {topic} for {audience}",
        f"Create a tutorial on {topic}",
        f"Ask questions about {topic} to engage {audience}",
        f"Share success stories related to {topic}",
        f"Provide tips and tricks for {topic}"
    ]


# Create the AI agent with all tools
ai_agent = Agent(
    tools=[
        generate_social_caption,
        suggest_hashtags,
        generate_image_with_nano_banana,
        generate_content_ideas
    ],
    name="Quflux Content Generator"
)


class AIContentService:
    """Service class for AI content generation operations."""
    
    def __init__(self):
        self.agent = ai_agent
    
    async def generate_caption(
        self,
        content_type: str,
        topic: str,
        platform: str,
        tone: str = "professional"
    ) -> str:
        """Generate a social media caption."""
        try:
            result = await self.agent.invoke_async(
                f"Generate a {tone} caption for {content_type} content about {topic} for {platform}"
            )
            return result
        except Exception as e:
            # Fallback to tool directly if agent fails
            return await generate_social_caption(content_type, topic, platform, tone)
    
    async def get_hashtag_suggestions(
        self,
        content: str,
        platform: str,
        max_hashtags: int = 10
    ) -> List[str]:
        """Get hashtag suggestions for content."""
        try:
            result = await self.agent.invoke_async(
                f"Suggest {max_hashtags} relevant hashtags for this {platform} content: {content}"
            )
            # Parse hashtags from result if needed
            return await suggest_hashtags(content, platform, max_hashtags)
        except Exception as e:
            return await suggest_hashtags(content, platform, max_hashtags)
    
    async def generate_image(
        self,
        prompt: str,
        style: str = "realistic",
        dimensions: str = "1024x1024"
    ) -> str:
        """Generate an image using AI."""
        try:
            result = await self.agent.invoke_async(
                f"Generate an image with this prompt: {prompt}, style: {style}, dimensions: {dimensions}"
            )
            return await generate_image_with_nano_banana(prompt, style, dimensions)
        except Exception as e:
            return await generate_image_with_nano_banana(prompt, style, dimensions)
    
    async def get_content_ideas(
        self,
        topic: str,
        audience: str,
        platform: str,
        count: int = 5
    ) -> List[str]:
        """Generate content ideas."""
        try:
            result = await self.agent.invoke_async(
                f"Generate {count} content ideas about {topic} for {audience} on {platform}"
            )
            return await generate_content_ideas(topic, audience, platform, count)
        except Exception as e:
            return await generate_content_ideas(topic, audience, platform, count)


# Global instance
ai_content_service = AIContentService()