backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
supabase>=2.3.0
celery>=5.3.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
boto3>=1.34.0
cryptography>=41.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.0
aiofiles>=23.2.0
pillow>=10.1.0
strands-agents
strands-agents-tools
sendgrid>=6.11.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
greenlet>=3.2.4